minor change that didn't fix anything but welp...i did something i guess

+++ io.openems.edge.bridge.genibus/src/io/openems/edge/bridge/genibus/api/GenibusTask.java
@@ -8,7 +8,7 @@ public interface GenibusTask {
 
     void setResponse(byte data);
 
-    int getAddress();
+    byte getAddress();
 
 
     int getHeader();
